app/requirements.txt
fastapi
uvicorn
requests
python-multipart
python-dotenv
pdf2image
Pillow
